# backend/src/main/java/com/vaultor/vaultor/controller/ResourceController.java
package com.vaultor.vaultor.controller;
import com.vaultor.vaultor.model.*;
import com.vaultor.vaultor.repository.*;
import com.vaultor.vaultor.service.*;
import com.vaultor.vaultor.controller.ApiDtos.*;
import lombok.RequiredArgsConstructor;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;
import org.springframework.http.*;
import org.springframework.core.io.UrlResource;
import java.util.*;

@RestController @RequestMapping("/api/resources") @RequiredArgsConstructor
public class ResourceController {
    private final ResourceRepository resources;
    private final ResourceBrowseService browse;
    private final ResourceSearchService searchService;
    private final RelationshipRepository relationships;
    private final ResourceService service;
    private final TagService tags;
    private final FileStorageService files;
    private final DocumentService documents;
    private final FileImportService imports;
    private final ImageService images;
    private final ResourceLifecycleService lifecycle;
    @GetMapping("/trash") public PageDto<ResourceSummary> trash(@RequestParam(defaultValue="0") int page,@RequestParam(defaultValue="100") int size,@RequestParam(defaultValue="") String q) {return browse.trash(page,size,q);}
    @PutMapping("/lifecycle") public List<ResourceLifecycleService.Result> lifecycle(@RequestBody List<ResourceLifecycleService.Item> items,@RequestHeader("X-Vaultor-Protocol") int protocol) {if(protocol<3)throw new org.springframework.web.server.ResponseStatusException(HttpStatus.UPGRADE_REQUIRED,"Update the client for safe resource lifecycle operations");return lifecycle.bulk(items);}
    @GetMapping("/{id}/usage") public ResourceLifecycleService.Usage usage(@PathVariable String id) {return lifecycle.usage(id);}
    @GetMapping("/{id}/lifecycle-pending") public ResourceLifecycleService.Item pending(@PathVariable String id) {return lifecycle.pending(id);}
    @PatchMapping("/{id}/title") public ResourceDto rename(@PathVariable String id,@RequestBody Map<String,String> input,@RequestHeader("If-Match") String expected) {return dto(service.rename(id,input.get("title"),expected));}
    @GetMapping("/{id}/image-info") public ImageService.Info imageInfo(@PathVariable String id) throws Exception { return images.info(id); }
    @PutMapping(value="/imports/{id}", consumes=MediaType.MULTIPART_FORM_DATA_VALUE)
    public ResourceDto importFile(@PathVariable String id, @RequestParam String title, @RequestParam(required=false) String content, @RequestParam(required=false) MultipartFile file, @RequestParam(required=false) String collectionId) throws Exception {
        return dto(imports.create(id,title,content,file,collectionId));
    }
    @GetMapping(value="/{id}/image-png", produces="image/png") public byte[] imagePng(@PathVariable String id) throws Exception { return images.clipboardPng(id); }
    private ResourceDto dto(Resource r) { return ResourceDto.of(r, documents); }
    private Resource get(String id) { return service.getResourceOrThrow(id); }
    private void validate(NoteInput input) {
        if (input.title() == null || input.title().isBlank() || input.title().length() > 500) throw new ApiErrors.FieldError("title", "Title must contain 1-500 characters");
        try { documents.validate(input.content()); } catch(IllegalArgumentException e) { throw new ApiErrors.FieldError("content", e.getMessage()); }
    }
    @GetMapping
    public PageDto<ResourceSummary> list(@RequestParam(defaultValue="0") int page, @RequestParam(defaultValue="100") int size,
        @RequestParam(defaultValue="") String q, @RequestParam(required=false) String type, @RequestParam(required=false) List<String> tag,
        @RequestParam(defaultValue="false") boolean favorites, @RequestParam(defaultValue="recent") String sort, @RequestParam(required=false) String collection) {
        return browse.list(page,size,q,type,tag == null ? List.of() : tag,favorites,sort,collection);
    }
    @GetMapping("/query") public PageDto<ResourceSearchService.Hit> query(@RequestParam String q,@RequestParam(defaultValue="0") int page,@RequestParam(defaultValue="50") int size,@RequestParam(required=false) String type,@RequestParam(required=false) List<String> tag,@RequestParam(required=false) String collection,@RequestParam(defaultValue="false") boolean favorites) {return searchService.search(q,page,size,type,tag==null?List.of():tag,collection,favorites);}
    @GetMapping("/search") public List<ResourceSummary> search(@RequestParam(defaultValue="") String q) {
        return browse.list(0,50,q,null,List.of(),false,"recent").items();
    }
    @PutMapping("/{id}/favorite") @ResponseStatus(HttpStatus.NO_CONTENT)
    public void favorite(@PathVariable String id, @RequestBody FavoriteInput input) { browse.favorite(id,input.favorite()); }
    @GetMapping("/{id}/summary") public ResourceSummary summary(@PathVariable String id) {var value=browse.byIds(List.of(id)).get(id);if(value==null)throw new NoSuchElementException("Resource not found");return value;}
    @GetMapping("/{id}") public ResourceDto one(@PathVariable String id) { return dto(get(id)); }
    @PostMapping("/{id}/open") public void open(@PathVariable String id) { browse.markOpened(id); }
    @PostMapping public ResourceDto create(@RequestBody NoteInput input) { validate(input); if (input.type()!=null && !input.type().equals("note")) throw new IllegalArgumentException("Use file upload for file resources"); return dto(service.createNote(input.title().trim(), input.content().toString(),input.collectionId())); }
    @PutMapping("/{id}/note") public ResourceDto update(@PathVariable String id, @RequestBody NoteInput input, @RequestHeader(value="If-Match", required=false) String expected,jakarta.servlet.http.HttpServletRequest request) { if(request.isSecure() && expected==null)throw new org.springframework.web.server.ResponseStatusException(HttpStatus.PRECONDITION_REQUIRED,"Network note updates require If-Match");validate(input); if (!get(id).getType().equals("note")) throw new IllegalArgumentException("Resource is not a note"); return dto(service.updateNote(id, input.title().trim(), input.content().toString(), expected)); }
    @PostMapping("/file") public ResourceDto upload(@RequestParam MultipartFile file) throws Exception { if(file.isEmpty()) throw new IllegalArgumentException("Choose a nonempty file"); return dto(service.uploadFile(file)); }
    @DeleteMapping("/{id}") @ResponseStatus(HttpStatus.NO_CONTENT) public void delete(@PathVariable String id) { get(id); service.deleteResource(id); }
    @GetMapping("/{id}/backlinks") public List<ResourceDto> backlinks(@PathVariable String id) { get(id); return resources.findAllById(relationships.findByToIdAndType(id,"link").stream().map(Relationship::getFromId).toList()).stream().filter(r->r.getTrashedAt()==null).map(this::dto).toList(); }
    @PostMapping("/{id}/replace-links") public void replace(@PathVariable String id, @RequestBody Map<String,String> input) { throw new org.springframework.web.server.ResponseStatusException(HttpStatus.METHOD_NOT_ALLOWED,"Combined link replacement and deletion is retired. Move originals to Trash; referring labels remain intact."); }
    @PostMapping("/{id}/tags/{name}") public void tag(@PathVariable String id,@PathVariable String name) { if(name.isBlank() || name.length()>100) throw new IllegalArgumentException("Tag must contain 1-100 characters"); tags.addTagToResource(id,name); }
    @DeleteMapping("/{id}/tags/{name}") public void untag(@PathVariable String id,@PathVariable String name) { tags.removeTagFromResource(id,name); }
    @GetMapping("/{id}/raw") public ResponseEntity<org.springframework.core.io.Resource> raw(@PathVariable String id) throws Exception { return binary(id,false); }
    @GetMapping("/{id}/download") public ResponseEntity<org.springframework.core.io.Resource> download(@PathVariable String id) throws Exception { return binary(id,true); }
    private ResponseEntity<org.springframework.core.io.Resource> binary(String id,boolean download) throws Exception {
        var r=get(id); if(!"file".equals(r.getType())) throw new IllegalArgumentException("Resource is not a file");
        var path=files.getFile(r.getFilePath()); if(!java.nio.file.Files.exists(path)) throw new NoSuchElementException("File bytes missing");
        MediaType mime; try { mime=MediaType.parseMediaType(r.getMimeType()); } catch(Exception e) { mime=MediaType.APPLICATION_OCTET_STREAM; }
        return ResponseEntity.ok().contentType(mime).header(HttpHeaders.CONTENT_DISPOSITION,(download ? ContentDisposition.attachment() : ContentDisposition.inline()).filename(r.getTitle(),java.nio.charset.StandardCharsets.UTF_8).build().toString()).body(new UrlResource(path.toUri()));
    }
}